requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
braceexpand==0.1.7
causal-conv1d==1.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
filelock==3.19.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
grpcio==1.74.0
hf-xet==1.1.8
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lightning-utilities==0.15.2
mamba-ssm==2.2.5
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
ninja==1.13.0
numpy==2.1.2
omegaconf==2.3.0
packaging==25.0
pillow==11.0.0
propcache==0.3.2
protobuf==6.32.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
setuptools==78.1.1
six==1.17.0
soundfile==0.13.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.19
tokenizers==0.21.4
torch==2.7.0
torchaudio==2.7.0
torchmetrics==1.8.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.55.4
triton==3.3.0
typing_extensions==4.15.0
urllib3==2.5.0
webdataset==1.0.2
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
